requirements.txt
# Build Tools
setuptools==68.0.0
wheel==0.40.0

# Base Server
fastapi==0.103.1
uvicorn[standard]==0.23.2
gunicorn==21.2.0

# Core Logic
pydantic==1.10.12
python-multipart==0.0.6
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7
passlib[bcrypt]==1.7.4

# Networking & Utils
requests
numpy==1.24.3
opencv-python-headless
fpdf2==2.7.4
